pyproject.toml
[tool.poetry]
name = "flux"
version = "0.1.0"
description = ""
authors = ["Samir Azzouz <s.azzouz97@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
nornir = "^3.0.0"
pynetbox = "^5.3.1"
napalm = "^3.2.0"
robotframework = "^3.2.2"
click = "^7.1.2"
jsnapy = "^1.3.5"
pybatfish = "^2020.12.23"
fastapi = "^0.63.0"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
